Renamed Kitten to Strigi
svn path=/trunk/playground/base/kitten/; revision=551675

+++ src/estraierindexer/estraierindexreader.h
@@ -18,12 +18,12 @@ private:
     static ESTCOND* createCondition(const jstreams::Query&);
     static const char* mapId(const std::string& id);
 public:
-    int countHits(const jstreams::Query&);
+    int32_t countHits(const jstreams::Query&);
     std::vector<jstreams::IndexedDocument> query(const jstreams::Query&);
     std::map<std::string, time_t> getFiles(char depth);
     int countDocuments();
-    int countWords();
-    int getIndexSize();
+    int32_t countWords();
+    int64_t getIndexSize();
 };
 
 #endif
