+++ src/luceneindexer/cluceneindexreader.h
@@ -34,12 +34,12 @@ private:
     static void addField(lucene::document::Field* field,
         jstreams::IndexedDocument&);
 public:
-    int countHits(const jstreams::Query&);
+    int32_t countHits(const jstreams::Query&);
     std::vector<jstreams::IndexedDocument> query(const jstreams::Query&);
     std::map<std::string, time_t> getFiles(char depth);
-    int countDocuments();
-    int countWords();
-    int getIndexSize();
+    int32_t countDocuments();
+    int32_t countWords();
+    int64_t getIndexSize();
 };
 
 #endif
