+++ src/streamindexer/indexreader.h
@@ -31,12 +31,12 @@ public:
 class IndexReader {
 public:
     virtual ~IndexReader() {}
-    virtual int countHits(const Query&) = 0;
+    virtual int32_t countHits(const Query&) = 0;
     virtual std::vector<IndexedDocument> query(const Query&) = 0;
     virtual std::map<std::string, time_t> getFiles(char depth) = 0;
-    virtual int countDocuments() { return -1; }
-    virtual int countWords() { return -1; }
-    virtual int getIndexSize() { return -1; }
+    virtual int32_t countDocuments() { return -1; }
+    virtual int32_t countWords() { return -1; }
+    virtual int64_t getIndexSize() { return -1; }
 };
 
 }
